+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [[v0.3.0]](https://github.com/mlange-42/arche-model/compare/v0.2.0...v0.3.0)
+
+### Features
+
+* Observer `RowToTable` as adapter from `Row` to `Table` observer (#37)
+* System `PerfTimer` prints total step and average time per tick on finalization (#37)
+
 ## [[v0.2.0]](https://github.com/mlange-42/arche-model/compare/v0.1.0...v0.2.0)
 
 ### Features